+++ packages/system/src/Name/api/hooks/document/index.ts
@@ -1,0 +1,13 @@
+import { useCustomers } from './useCustomers';
+import { useSuppliers } from './useSuppliers';
+import { useInternalSuppliers } from './useInternalSuppliers';
+import { useName } from './useName';
+import { useNames } from './useNames';
+
+export const Document = {
+  useCustomers,
+  useSuppliers,
+  useInternalSuppliers,
+  useName,
+  useNames,
+};
