+++ packages/system/src/Name/api/hooks/index.ts
@@ -1,0 +1,13 @@
+import { Document } from './document';
+import { Utils } from './utils';
+
+export const useName = {
+  api: Utils.useNameApi,
+  document: {
+    get: Document.useName,
+    customers: Document.useCustomers,
+    internalSuppliers: Document.useInternalSuppliers,
+    list: Document.useNames,
+    suppliers: Document.useSuppliers,
+  },
+};
