+++ packages/system/src/Name/api/hooks/utils/index.ts
@@ -1,0 +1,5 @@
+import { useNameApi } from './useNameApi';
+
+export const Utils = {
+  useNameApi,
+};
